refactor: isolate e2e utility setup
