+++ e2e/helpers.ts
@@ -75,7 +75,7 @@ export async function loadAudioFile(
  * debounce.
  */
 export async function evaluateFlushAutoSave(page: Page): Promise<void> {
-  await page.evaluate(() => window.__e2e!.flushAutoSave());
+  await page.evaluate(() => window.__e2e.flushAutoSave());
 }
 
 /**
@@ -96,7 +96,7 @@ export async function evaluateStore<T>(
   fn: (store: typeof useProjectStore) => T,
 ): Promise<T> {
   return page.evaluate((fnStr) => {
-    const store = window.__e2e!.useProjectStore;
+    const store = window.__e2e.useProjectStore;
     const evalFn = new Function("store", `return (${fnStr})(store)`);
     return evalFn(store) as T;
   }, fn.toString());
