+++ e2e/project-migration.spec.ts
@@ -44,7 +44,7 @@ test.describe("Project Migration", () => {
     const audio = new Uint8Array(await readFile(TEST_AUDIO_PATH));
     await page.evaluate(
       async ({ project, audio }) => {
-        await window.__e2e!.seedProjectV1("Legacy Project", project, audio);
+        await window.__e2e.seedProjectV1("Legacy Project", project, audio);
       },
       { project: LEGACY_PROJECT, audio },
     );
